basic_novel_scrapper_flask_app/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
beautifulsoup4
python-dotenv
cloudscraper